Improve Makefile (multiple fixes)
DESTDIR is usually set from outside (eg. when packaging applications) to build inside a destination, this replaces DESTDIR with PREFIX while preserving the DESTDIR usage on the paths, allowing for external definition without unexpected behavior.
This patch also replaces the absolute link in the install section with a relative link, thus making the files more portable.